Changed TileRows to TileRow Moved TableBody and departures map to DepartureTile
